requirements.txt
# Telegram Bot
python-telegram-bot[webhooks]==21.6

# Web framework
fastapi==0.115.4
uvicorn[standard]==0.32.0

# Database
asyncpg==0.30.0
SQLAlchemy[asyncio]==2.0.36
alembic==1.14.0

# AI
openai==1.54.4

# Scheduler
apscheduler==3.10.4

# Auth / Security
python-jose[cryptography]==3.3.0
passlib==1.7.4

# Config / Validation
pydantic==2.9.2
pydantic-settings==2.6.1

# HTTP client
httpx==0.27.2

# AWS SDK (Secrets Manager, CodeDeploy)
boto3==1.35.67

# Utilities
python-multipart==0.0.12
pytz==2024.2
aiofiles==24.1.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2

# Dev
mypy==1.13.0
black==24.10.0
isort==5.13.2
